java json-ast: extract Node to ast
