+++ tools/json-ast/x/java/ast.go
@@ -1,0 +1,32 @@
+package java
+
+import sitter "github.com/smacker/go-tree-sitter"
+
+// Node models a tree-sitter node for Java source.
+type Node struct {
+	Kind     string `json:"kind"`
+	Start    int    `json:"start"`
+	End      int    `json:"end"`
+	Text     string `json:"text,omitempty"`
+	Children []Node `json:"children,omitempty"`
+}
+
+// convert creates a Node from a tree-sitter node.
+func convert(n *sitter.Node, src []byte) Node {
+	node := Node{
+		Kind:  n.Type(),
+		Start: int(n.StartByte()),
+		End:   int(n.EndByte()),
+	}
+	if n.ChildCount() == 0 {
+		node.Text = n.Content(src)
+	}
+	for i := 0; i < int(n.ChildCount()); i++ {
+		child := n.Child(i)
+		if child == nil {
+			continue
+		}
+		node.Children = append(node.Children, convert(child, src))
+	}
+	return node
+}
